# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pearlis Luxury Jewelry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # REVIEWS
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /cart/coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      summary: Apply coupon to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyCouponBody"
      responses:
        "200":
          description: Discount info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResult"

  # ORDERS
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List blog posts
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 9 }
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
    post:
      operationId: createBlog
      tags: [blogs]
      summary: Create blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
  /blogs/{id}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get blog post
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    put:
      operationId: updateBlog
      tags: [blogs]
      summary: Update blog (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    delete:
      operationId: deleteBlog
      tags: [blogs]
      summary: Delete blog (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # COUPONS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    Product:
      type: object
      required: [id, name, price, category, images, stock, isNew, isTrending, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        category:
          type: string
        categoryId:
          type: integer
        material:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isNew:
          type: boolean
        isTrending:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        productCount:
          type: integer
    CreateCategoryBody:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
...
    CartItem:
      type: object
      required: [productId, quantity, product]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
    Cart:
      type: object
      required: [items, subtotal, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        couponCode:
          type: string
          nullable: true
    AddToCartBody:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    UpdateCartItemBody:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    ApplyCouponBody:
      type: object
      required: [code]
      properties:
        code:
          type: string
    CouponResult:
      type: object
      required: [valid, discount]
      properties:
        valid:
          type: boolean
        discount:
          type: number
        message:
          type: string
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatar:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    RegisterBody:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
          nullable: true
...
    Blog:
      type: object
      required: [id, title, slug, excerpt, content, imageUrl, author, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    BlogListResponse:
      type: object
      required: [blogs, total, page, limit]
      properties:
        blogs:
          type: array
          items:
            $ref: "#/components/schemas/Blog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateBlogBody:
      type: object
      required: [title, excerpt, content, imageUrl, author]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string
...